DBus and gnome-keyring fixed.
PROBLEM:
The issue  was that the  process `xbindkeys` was launched  outside the
DBUS session.  Since almost all  programs where started either  from a
keybind or from a terminal launched by `xkeybinds`, almost no apps had
access to the DBUS appropriate variables.

CAUSE:
This was a bug of `xbindkeys` xsession scripts.

FIX:
Do not install  an `xbindkeysrc` file in the home  folder, in order to
inhibit the `xbindkeys` to be launched erratically by gdm.
Process `xbindkeys` must be launched manually from the `xsession`
script, in order to give it the appropriate environment variables.